Implement Autonomous Zero-Click Agentic Onboarding Engine
This PR resolves #30882 by migrating the "Zero-Click Business Generator" from the growth endpoints to the onboarding module. The endpoint is moved to `/api/v1/onboarding/generate` and all UI/e2e integration code is updated to route towards the new API.

Changes:
- Refactored `handle_zero_click_generate` out of `src/server/api/growth.rs` and moved it into `src/server/api/onboarding/mod.rs` (with test scenarios updated to test `OnboardingAgent` rather than `GrowthState`).
- Updated `src/server/lib.rs` nesting to export `api::onboarding::router` at `/api/v1/onboarding`.
- Updated all references in `src/ui/next/src/app/zero-click-builder/page.tsx`, `src/e2e/ui/zero_click_builder.spec.ts`, `src/ui/tauri/src/ui/setup.html` and `src/ui/tauri/src/ui/index.html` to consume the new `api/v1/onboarding/generate` endpoint.
- Ensured 100% Hermetic Testing passing locally and within all `src/server` and Playwright contexts, with strict verification checks.

Resolves #30882
