+++ src/e2e/ui/zero_click_builder.spec.ts
@@ -30,7 +30,7 @@ test.describe('Zero-Click Business Generator CUJ', () => {
     });
 
     // Mock the api response
-    await page.route('**/api/v1/growth/zero-click-builder/generate*', async route => {
+    await page.route('**/api/v1/onboarding/generate*', async route => {
         await route.fulfill({
             status: 200,
             contentType: 'application/json',
