+++ src/ui/tauri/src/ui/index.html
@@ -160,7 +160,7 @@
 
           try {
              const backendUrl = (window.location.protocol === 'file:' || ["1420", "3000", "8081"].includes(window.location.port)) ? 'http://127.0.0.1:18789' : '';
-             const res = await fetch(`${backendUrl}/api/v1/growth/zero-click-builder/generate`, {
+             const res = await fetch(`${backendUrl}/api/v1/onboarding/generate`, {
                  method: 'POST',
                  headers: {
                      'Content-Type': 'application/json',
