+++ src/ui/tauri/src/ui/setup.html
@@ -2385,7 +2385,7 @@ if (state.capabilities && document.getElementById('cap-draft')) {
 
               try {
                   const backendUrl = (window.location.protocol === 'file:' || ["1420", "3000", "8081"].includes(window.location.port)) ? 'http://127.0.0.1:18789' : '';
-                  const res = await fetch(`${backendUrl}/api/v1/growth/zero-click-builder/generate`, {
+                  const res = await fetch(`${backendUrl}/api/v1/onboarding/generate`, {
                       method: 'POST',
                       headers: {
                           'Content-Type': 'application/json',
